fixed eTransfer test
