+++ libs/rbcapi/src/index.ts
@@ -1,8 +1,7 @@
 import { log } from '@thecointech/logging';
-import { readFileSync } from 'fs';
 import { RbcStore } from './store';
 import { ApiAction } from './action';
-import { ProgressCallback, isCredentials, AuthOptions } from './types';
+import { ProgressCallback, AuthOptions } from './types';
 
 import { getTransactions, fetchLatestTransactions } from './transactions';
 import { depositETransfer } from './deposit';
@@ -35,25 +34,7 @@ export class RbcApi {
   }
 
   initialize(options?: AuthOptions) {
-    if (isCredentials(options))
-      ApiAction.Credentials = options;
-    else if (options?.authFile) {
-      // it's a json file
-      const cred = readFileSync(options.authFile, 'utf8');
-      ApiAction.Credentials = JSON.parse(cred);
-    }
-    else if (process.env.RBCAPI_CREDENTIALS) {
-      // Use environment vars if available
-      ApiAction.Credentials = JSON.parse(process.env.RBCAPI_CREDENTIALS);
-    }
-    else if (process.env.RBCAPI_CREDENTIALS_PATH) {
-      // Use environment vars if available
-      const cred = readFileSync(process.env.RBCAPI_CREDENTIALS_PATH, 'utf8');
-      ApiAction.Credentials = JSON.parse(cred);
-    }
-    else {
-      throw new Error('Cannot use RbcApi without credentials');
-    }
+    ApiAction.initCredentials(options);
     RbcStore.Options = ApiAction.Credentials.timezone;
   }
 }
